refactor: deleting unnecessary widgets
