+++ pubspec.yaml
@@ -35,11 +35,11 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  
   hooks_riverpod: ^1.0.4
   currency_formatter: ^1.2.1
   intl: ^0.17.0
-  money_formatter: ^0.0.3
+ 
 
 dev_dependencies:
   flutter_test:
